consolidated and fixed Registration
